update lesson 5

+++ app/travel_together/service.py
@@ -1,14 +1,20 @@
 import uuid
 
-from app.config import Settings
-from app.travel_together.models.participant import ParticipanStatus
+from app.travel_together.permissions import PermissionService
+from app.travel_together.models import ParticipantStatus
+from app.config import get_settings
 from app.exceptions import (
-    TripOrganizerRequiredError,
     AlreadyTripParticipant,
     ReachedMaxParticipants,
+    TripOrganizerRequiredError,
+    TripNotFoundError,
 )
 from .repository import TripRepository, ParticipantRepository
-from .schemas import TripCreate, TripResponse, ParticipantResponse
+from .schemas import (
+    TripCreate,
+    TripResponse,
+    ParticipantResponse,
+)
 
 
 class TripService:
@@ -16,27 +22,17 @@ class TripService:
         self,
         trip_repo: TripRepository,
         participant_repo: ParticipantRepository,
+        permission_service: PermissionService,
     ):
-
         self.trip_repo = trip_repo
         self.participant_repo = participant_repo
+        self.permission_service = permission_service
 
-    async def create_trip(
-        self, trip: TripCreate, current_user_id: uuid.UUID
-    ) -> TripResponse:
+    async def create_trip(self, trip: TripCreate, user_id: uuid.UUID) -> TripResponse:
         trip_data = trip.model_dump()
-        trip_data["organizer_id"] = current_user_id
+        trip_data["organizer_id"] = user_id
         trip = await self.trip_repo.create_trip(trip_data)
 
-        existing = await self.participant_repo.check_participant_exists(
-            trip_id=trip.id, user_id=current_user_id
-        )
-        if not existing:
-            await self.participant_repo.add_participant(
-                trip_id=trip.id,
-                current_user_id=current_user_id,
-                status=ParticipanStatus.ACCEPTED,
-            )
         return TripResponse.model_validate(trip)
 
     async def retrieve_trip(self, trip_id: uuid.UUID) -> TripResponse:
@@ -44,26 +40,28 @@ class TripService:
         return TripResponse.model_validate(trip)
 
     async def update_trip(
-        self, trip_id: uuid.UUID, trip: TripCreate, current_user_id: uuid.UUID
+        self, trip_id: uuid.UUID, trip: TripCreate, user_id: uuid.UUID
     ) -> TripResponse:
         existing_trip = await self.trip_repo.retrieve_trip(trip_id)
 
-        if existing_trip.organizer_id != current_user_id:
-            raise TripOrganizerRequiredError("Only the organaizer can delete trip")
+        if not await self.permission_service.check_is_user_trip_organizer(
+            user_id, trip_id
+        ):
+            raise TripOrganizerRequiredError("Only the organizer can update this trip.")
 
         trip_data = trip.model_dump()
-        # не меняем организатора + явно его сохраняем
         trip_data["organizer_id"] = existing_trip.organizer_id
         updated_trip = await self.trip_repo.update_trip(trip_id, trip_data)
 
         return TripResponse.model_validate(updated_trip)
 
-    async def delete_trip(self, trip_id: uuid.UUID, current_user_id: uuid.UUID) -> None:
-        trip = await self.trip_repo.retrieve_trip(trip_id)
+    async def delete_trip(self, trip_id: uuid.UUID, user_id: uuid.UUID) -> None:
+        if not await self.permission_service.check_is_user_trip_organizer(
+            user_id, trip_id
+        ):
+            raise TripOrganizerRequiredError("Only the organizer can delete this trip.")
 
-        if trip.organizer_id != current_user_id:
-            raise TripOrganizerRequiredError("Only the organaizer can update trip")
-        return await self.trip_repo.delete_trip(trip_id)
+        await self.trip_repo.delete_trip(trip_id)
 
 
 class ParticipantService:
@@ -71,10 +69,12 @@ class ParticipantService:
         self,
         trip_repo: TripRepository,
         participant_repo: ParticipantRepository,
+        permission_service: PermissionService,
     ):
         self.trip_repo = trip_repo
         self.participant_repo = participant_repo
-        self.settings = Settings()
+        self.permission_service = permission_service
+        self.settings = get_settings()
 
     async def add_participant(
         self, trip_id: uuid.UUID, user_id: uuid.UUID
@@ -82,8 +82,7 @@ class ParticipantService:
         # Используем retrieve_trip_for_update для блокировки строки
         # чтобы избежать состояние гонки
         trip = await self.trip_repo.retrieve_trip_for_update(trip_id)
-
-        if trip.organizer_id == user_id:
+        if await self.permission_service.check_is_user_trip_organizer(user_id, trip.id):
             raise AlreadyTripParticipant(
                 "Organizer is already a participant with ACCEPTED status."
             )
@@ -96,16 +95,31 @@ class ParticipantService:
             raise ReachedMaxParticipants("Reached max amount of participants")
 
         result = await self.participant_repo.add_participant(
-            trip_id=trip_id, current_user_id=user_id, status=ParticipanStatus.PENDING
+            trip_id=trip_id, user_id=user_id, status=ParticipantStatus.PENDING
         )
 
         return ParticipantResponse.model_validate(result)
 
-    async def remove_participant(self, trip_id: uuid.UUID, user_id: uuid.UUID) -> bool:
-        return await self.participant_repo.remove_participant(trip_id, user_id)
-
-    async def retrieve_participants(
+    async def retrieve_all_participants_from_trip(
         self, trip_id: uuid.UUID
     ) -> list[ParticipantResponse]:
-        items = await self.participant_repo.retrieve_participants(trip_id)
+        items = await self.participant_repo.retrieve_all_participants_from_trip(trip_id)
         return [ParticipantResponse.model_validate(item) for item in items]
+
+    async def remove_participant(self, trip_id: uuid.UUID, user_id: uuid.UUID) -> bool:
+        trip = await self.trip_repo.retrieve_trip(trip_id)
+        if not trip:
+            raise TripNotFoundError(trip_id)
+        is_organizer = await self.permission_service.check_is_user_trip_organizer(
+            user_id, trip_id
+        )
+        is_participant_exists = (
+            await self.permission_service.check_is_user_trip_participant(
+                user_id, trip_id
+            )
+        )
+        if not (is_organizer or is_participant_exists):
+            raise TripOrganizerRequiredError(
+                "Only the trip organizer or trip participant can remove trip member"
+            )
+        return await self.participant_repo.remove_participant(trip_id, user_id)
